+++ joulescope_ui/plugins/frequency.py
@@ -123,7 +123,7 @@ class FrequencyRangeToolWidget(QtWidgets.QWidget):
             'dtype': 'obj',
             'brief': 'Hold the histogram data',
             'default': None,
-            'flags': ['hidden'],
+            'flags': ['hide'],
         }
     }